agregada ruta para eliminar y actualizar periodo

+++ app/Http/Controllers/Periodos/DeleteProcessController.php
@@ -10,6 +10,11 @@ class DeleteProcessController extends Controller
     public function __invoke($id)
     {
         $periodo = Periodo::findOrFail($id);
+
+        if ($periodo->encuestasRelacionadas()) {
+            return response()->json(['message' => 'El periodo tiene encuestas relacionadas.'], 409);
+        }
+
         $periodo->delete();
 
         return response()->json();
